[FLINK-14553][runtime] Respect non-blocking output in StreamTask#processInput
The non-blocking output was introduced in FLINK-14396 and FLINK-14498 to solve the problem of handling the checkpoint barrier in the case of backpressure.
In order to make the whole process through, StreamInputProcessor should be allowed to process input elements if the output is also available.
The default core size of LocalBufferPool for ResultPartition should also be increased by 1 in order not to impact the performance in the new way, and this tiny memory overhead could be ignored in practice.